refactor: separate Feed and FeedSource update logic
- Move Feed.update() to orchestrate both FeedSource and Feed updates
- Create FeedSource.update() to handle source metadata updates only
- Add Feed.update(for:) to handle feed item updates per source
- Improves separation of concerns and maintainability
