+++ General/Models/Feed/Feed.swift
@@ -8,6 +8,7 @@
 import FeedKit
 import Foundation
 import SwiftData
+import SwiftUI
 
 @Model
 final class Feed {
@@ -20,6 +21,13 @@ final class Feed {
     @Attribute(.unique) var url: URL
     var imageURL: URL?
     var colorHex: String?
+
+    var color: Color {
+        if let hex = colorHex {
+            return Color(hex: hex)
+        }
+        return source?.color ?? Color.fromSeed(title)
+    }
 
     init(
         title: String,
@@ -41,6 +49,110 @@ final class Feed {
 }
 
 extension Feed {
+    @MainActor
+    static func update() async throws {
+        try await FeedSource.update()
+
+        let descriptor = FetchDescriptor<FeedSource>()
+        let feedSources = try SwiftDataStack.modelContext.fetch(descriptor)
+
+        for feedSource in feedSources {
+            try await Feed.update(for: feedSource)
+        }
+    }
+
+    @MainActor
+    static func update(for feedSource: FeedSource) async throws {
+        let parseResult = try? await withCheckedThrowingContinuation { continuation in
+            FeedParser(URL: feedSource.url)
+                .parseAsync { result in
+                    switch result {
+                    case .success(let success):
+                        continuation.resume(returning: success)
+                    case .failure(let failure):
+                        continuation.resume(throwing: failure)
+                    }
+                }
+        }
+
+        if let rssItems = parseResult?.rssFeed?.items {
+            for item in rssItems {
+                let feedURL = URL(string: item.link ?? "example.com")!
+                let feedTitle = item.title ?? "!!No title found for this Feed"
+                let feedKeywords = Set(item.categories?.map { $0.value ?? "" } ?? [])
+                let feedDescription = item.description
+                let feedDate = item.pubDate ?? Date()
+                let feedColorHex = feedSource.colorHex
+
+                var feedImageURL: URL?
+                if let enclosure = item.enclosure, enclosure.attributes?.type == "image/jpeg",
+                    let urlString = enclosure.attributes?.url
+                {
+                    feedImageURL = URL(string: urlString)
+                }
+
+                try SwiftDataStack.modelContext.upsert(
+                    predicate: #Predicate<Feed> { $0.url == feedURL },
+                    update: { feed in
+                        feed.title = feedTitle
+                        feed.keywords = feedKeywords
+                        feed.detailText = feedDescription
+                        feed.datePosted = feedDate
+                        feed.imageURL = feedImageURL
+                        feed.colorHex = feedColorHex
+                        feed.source = feedSource
+                    },
+                    create: {
+                        let newFeed = Feed(
+                            title: feedTitle,
+                            keywords: feedKeywords,
+                            detailText: feedDescription,
+                            datePosted: feedDate,
+                            url: feedURL,
+                            imageURL: feedImageURL,
+                            colorHex: feedColorHex
+                        )
+                        newFeed.source = feedSource
+                        return newFeed
+                    }
+                )
+            }
+        } else if let atomEntries = parseResult?.atomFeed?.entries {
+            for entry in atomEntries {
+                let feedURL = URL(string: entry.links?.first?.attributes?.href ?? "example.com")!
+                let feedTitle = entry.title ?? "!!No title found for this Feed"
+                let feedKeywords = Set(entry.categories?.map { $0.attributes?.label ?? "" } ?? [])
+                let feedDescription = entry.summary?.value
+                let feedDate = entry.updated ?? Date()
+                let feedColorHex = feedSource.colorHex
+
+                try SwiftDataStack.modelContext.upsert(
+                    predicate: #Predicate<Feed> { $0.url == feedURL },
+                    update: { feed in
+                        feed.title = feedTitle
+                        feed.keywords = feedKeywords
+                        feed.detailText = feedDescription
+                        feed.datePosted = feedDate
+                        feed.colorHex = feedColorHex
+                        feed.source = feedSource
+                    },
+                    create: {
+                        let newFeed = Feed(
+                            title: feedTitle,
+                            keywords: feedKeywords,
+                            detailText: feedDescription,
+                            datePosted: feedDate,
+                            url: feedURL,
+                            colorHex: feedColorHex
+                        )
+                        newFeed.source = feedSource
+                        return newFeed
+                    }
+                )
+            }
+        }
+    }
+
     convenience init(item: RSSFeedItem, source: FeedSource) {
         self.init(
             title: item.title ?? "!!No title found for this Feed",
